Separate ui and non-ui related files
