+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from 'app/App';
+import App from 'ui/app/App';
 
 import 'bootstrap/dist/css/bootstrap.css';
 import 'index.css';
