+++ src/ui/app/App.test.js
@@ -1,9 +1,10 @@
-import App from 'app/App';
-import AppRoutes from 'app/AppRoutes';
-import AuthenticationIndicator from 'authentication/AuthenticationIndicator';
-
 import React from 'react';
 import { shallow } from 'enzyme';
+
+import App from 'ui/app/App';
+import AppRoutes from 'ui/app/AppRoutes';
+import AuthenticationIndicator from 'ui/auth/AuthenticationIndicator';
+
 
 describe('App', () => {
     it('renders AuthenticationIndicator while the user has not yet authenticated', () => {
