+++ src/ui/app/AppRoutes.js
@@ -2,8 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
-import BudgetSummary from 'summary/BudgetSummaryPage';
-import Transactions from 'transaction/TransactionsPage';
+import BudgetSummary from 'ui/summary/BudgetSummaryPage';
+import Transactions from 'ui/transaction/TransactionsPage';
+
 
 export default class AppRoutes extends React.Component {
     render() {
