+++ src/ui/auth/AuthenticationIndicator.css
