+++ src/ui/format/CurrencyFormat.js
